requirements.txt
python-dotenv
requests
dash
pandas
matplotlib

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

paho-mqtt
